+++ CHANGELOG.rst
@@ -1,9 +1,41 @@
-// Release notes
-// -------------
+Changelog
+=========
 
-### Unreleased
+Unreleased
+----------
+
+- Fix a server error on resource details view when the compliance alert is "missing".
+  https://github.com/nexB/scancode.io/issues/344
+
+- Migrate the ScanCodebase pipeline from `scancode.run_scancode` subprocess to
+  `scancode.scan_for_application_packages` and `scancode.scan_for_files`.
+  https://github.com/nexB/scancode.io/issues/340
+
+v30.0.1 (2021-10-11)
+--------------------
+
+- Fix a build failure related to dependency conflict.
+  https://github.com/nexB/scancode.io/issues/342
+
+v30.0.0 (2021-10-8)
+-------------------
 
 - Upgrade ScanCode-toolkit to version 30.1.0
+
+- Replace the task queue system, from Celery to RQ.
+  https://github.com/nexB/scancode.io/issues/176
+
+- Add ability to delete "not started" and "queued" pipeline tasks.
+  https://github.com/nexB/scancode.io/issues/176
+
+- Add ability to stop "running" pipeline tasks.
+  https://github.com/nexB/scancode.io/issues/176
+
+- Refactor the "execute" management command and add support for --async mode.
+  https://github.com/nexB/scancode.io/issues/130
+
+- Include codebase resource data in the details of package creation project errors.
+  https://github.com/nexB/scancode.io/issues/208
 
 - Add a SCANCODEIO_REST_API_PAGE_SIZE setting to control the number of objects
   returned per page in the REST API.
@@ -12,7 +44,8 @@
 - Provide an "add input" action on the Project endpoint of the REST API.
   https://github.com/nexB/scancode.io/issues/318
 
-### v21.9.6
+v21.9.6
+-------
 
 - Add ability to "archive" projects, from the Web UI, API and command line interface.
   Data cleanup of the project's input, codebase, and output directories is available
@@ -37,7 +70,8 @@
 - Log the outputs of run_scancode as progress indication.
   https://github.com/nexB/scancode.io/issues/300
 
-### v21.8.2
+v21.8.2
+-------
 
 - Upgrade ScanCode-toolkit to version 21.7.30
 
@@ -66,7 +100,8 @@
 - Add support for VMWare Photon-based Docker images and rootfs. This is an RPM-based
   Linux distribution
 
-### v21.6.10
+v21.6.10
+--------
 
 - Add support for VM image formats extraction such as VMDK, VDI and QCOW.
   See https://github.com/nexB/extractcode#archive-format-kind-file_system for the full
@@ -91,7 +126,8 @@
 - Add the is_binary, is_text, and is_archive fields to the CodebaseResource model.
   https://github.com/nexB/scancode.io/issues/75
 
-### v21.5.12
+v21.5.12
+--------
 
 - Adds a new way to fetch docker images using skopeo provided as a
   plugin using docker:// reference URL-like pointers to a docker image.
@@ -107,7 +143,8 @@
   load_inventory pipelines.
   https://github.com/nexB/scancode.io/issues/147
 
-### v21.4.28
+v21.4.28
+--------
 
 - The installation local timezone can be configured using the TIME_ZONE setting.
   The current timezone in now included in the dates representation in the web UI.
@@ -140,7 +177,8 @@
 
 - Add links to codebase resource and to discovered packages in list views.
 
-### v21.4.14
+v21.4.14
+--------
 
 - Implement timeout on the scan functions, default to 120 seconds per resources.
   https://github.com/nexB/scancode.io/issues/135
@@ -149,7 +187,8 @@
   https://github.com/nexB/scancode.io/issues/116
   https://github.com/nexB/scancode.io/issues/141
 
-### v21.4.5
+v21.4.5
+-------
 
 - Add support for Docker and VM images using RPMs such as Fedora, CentOS, RHEL,
   and openSUSE linux distributions.
@@ -192,7 +231,8 @@
 
 - Consistent ordering for the pipelines, now sorted alphabetically.
 
-### v1.1.0 (2021-02-16)
+v1.1.0 (2021-02-16)
+-------------------
 
 - Display project extra data in the project details view.
   https://github.com/nexB/scancode.io/issues/88
@@ -231,7 +271,8 @@
   WARNING: The new Pipelines syntax is not backward compatible with v1.0.x
   https://github.com/nexB/scancode.io/issues/82
 
-### v1.0.7 (2021-02-01)
+v1.0.7 (2021-02-01)
+-------------------
 
 - Add user interface to manage Projects from a web browser
   All the command-line features are available
@@ -245,7 +286,8 @@
 - Add a profile() method on the Run model for profiling pipeline execution
   https://github.com/nexB/scancode.io/issues/73
 
-### v1.0.6 (2020-12-23)
+v1.0.6 (2020-12-23)
+-------------------
 
 - Add a management command to delete a Project and its related work directories
   https://github.com/nexB/scancode.io/issues/65
@@ -268,7 +310,8 @@
 - Install psycopg2-binary instead of psycopg2 on non-Linux platforms
   https://github.com/nexB/scancode.io/issues/64
 
-### v1.0.5 (2020-12-07)
+v1.0.5 (2020-12-07)
+-------------------
 
 - Add minimal license list and text views
   https://github.com/nexB/scancode.io/issues/32
@@ -293,7 +336,8 @@
 - Add support for Python 3.7, 3.8, and 3.9
   https://github.com/nexB/scancode.io/issues/54
 
-### v1.0.4 (2020-11-17)
+v1.0.4 (2020-11-17)
+-------------------
 
 - Add a to_json output pipe returning ScanCode compatible content
   https://github.com/nexB/scancode.io/issues/45
@@ -310,7 +354,8 @@
 
 - Upgrade Django, Metaflow, and ScanCode-toolkit to latest versions
 
-### v1.0.3 (2020-09-24)
+v1.0.3 (2020-09-24)
+-------------------
 
 - Add ability to resume a failed pipeline from the run management command
   Fix for https://github.com/nexB/scancode.io/issues/22
@@ -321,7 +366,8 @@
 - Add support for "failed" task_output in Run.get_run_id method
   Fix for https://github.com/nexB/scancode.io/issues/17
 
-### v1.0.2 (2020-09-18)
+v1.0.2 (2020-09-18)
+-------------------
 
 - Add documentation and tutorial
   For https://github.com/nexB/scancode.io/issues/8
@@ -333,13 +379,15 @@
 - Always return the Pipeline subclass/implementation from the module inspection
   Fix for https://github.com/nexB/scancode.io/issues/11
 
-### v1.0.1 (2020-09-12)
+v1.0.1 (2020-09-12)
+-------------------
 
 - Do not fail when collecting system packages in Ubuntu docker images for
   layers that do not install packages by updating to a newer version of
   ScanCode Toolkit
   Fix for https://github.com/nexB/scancode.io/issues/1
 
-### v1.0.0 (2020-09-09)
+v1.0.0 (2020-09-09)
+-------------------
 
 - Initial release